Add experiment module
experiment.rs pertains to running many replications of the same
Simulation.

There are a few requirements this module might serve:

1. Sometimes, a researcher may want to run a simulation with the same
parameters many times to increase the precision of the confidence
interval for the results of the simulation.

2. A researcher may want to run many replications of a simulation over a
wide domain of parameter variations. This simulated annealing
functionality, to approximate a global optimization in a large search
space for an optimization problem is also contained here.